+++ main.go
@@ -123,6 +123,31 @@ func main() {
            to each path. Both keys will be PEM-encoded DER. (nbits defaults
            to 2048 bits)
 
+    cert role path
+           Generates a signed Certificate using Vault's PKI backend + Certifiate
+           Authority using the provided role. The common name is derived from the
+           last part of the path provided. Once issued, safe will store the
+           private key, signed certificate, and serial number in the secret backend,
+           located at the specified path.
+
+           The --ttl, --ip-sans, --alt-names, and --exclude-cn-from-sans flags can
+           be specified to customize how the certificate is generated.
+
+
+    revoke path|serial
+           Revokes a certificate from Vaults PKI backend, using the specified serial,
+           or path to a secret containing the serial of the certificate. Once revoked,
+           the CRL will be automatically updated inside Vault, but anything consuming
+           the CRL should pull a new copy.
+
+    ca-pem
+           Retrieves the PEM-encoded CA cert used in Vault's PKI backend for signing
+           and issuing certificates.
+
+    crl-pem
+           Retrieves the PEM-encoded Certificate Revocation List managed by
+           Vaults PKI backend.
+
     prompt ...
            Echo the arguments, space-separated, as a single line to the terminal.
 
@@ -607,6 +632,88 @@ func main() {
 		}
 
 		return v.Write(path, s)
+	})
+
+	r.Dispatch("crl-pem", func(command string, args ...string) error {
+		rc.Apply()
+
+		v := connect()
+		pem, err := v.RetrievePem("crl")
+		if err != nil {
+			return err
+		}
+		if len(pem) == 0 {
+			ansi.Fprintf(os.Stderr, "@Y{No CRL exists yet}\n")
+		} else {
+			fmt.Fprintf(os.Stdout, "%s\n", pem)
+		}
+		return nil
+	})
+
+	r.Dispatch("ca-pem", func(command string, args ...string) error {
+		rc.Apply()
+
+		v := connect()
+		pem, err := v.RetrievePem("ca")
+		if err != nil {
+			return err
+		}
+		if len(pem) == 0 {
+			ansi.Fprintf(os.Stderr, "@Y{No CA exists yet}\n")
+		} else {
+			fmt.Fprintf(os.Stdout, "%s\n", pem)
+		}
+		return nil
+	})
+
+	r.Dispatch("cert", func(command string, args ...string) error {
+		rc.Apply()
+
+		ttl := getopt.StringLong("ttl", 0, "", "Vault-compatible time specification for the length the Cert is valid for")
+		ip_sans := getopt.StringLong("ip-sans", 0, "", "Comma-separated list of IP SANs")
+		alt_names := getopt.StringLong("alt-names", 0, "", "Comma-separated list of SANs")
+		exclude_cn_from_sans := getopt.BoolLong("exclude-cn-from-sans", 0, "", "Exclude the common_name from DNS or Email SANs")
+
+		args = append([]string{"safe " + command}, args...)
+
+		var opts = getopt.CommandLine
+		var parsed []string
+		for {
+			opts.Parse(args)
+			if opts.NArgs() == 0 {
+				break
+			}
+			parsed = append(parsed, opts.Arg(0))
+			args = opts.Args()
+		}
+
+		args = parsed
+
+		params := vault.CertOptions{
+			TTL:               *ttl,
+			IPSans:            *ip_sans,
+			AltNames:          *alt_names,
+			ExcludeCNFromSans: *exclude_cn_from_sans,
+		}
+
+		if len(args) != 2 {
+			return fmt.Errorf("USAGE: cert role path")
+		}
+
+		v := connect()
+		role, path := args[0], args[1]
+		return v.CreateSignedCertificate(role, path, params)
+	})
+
+	r.Dispatch("revoke", func(command string, args ...string) error {
+		rc.Apply()
+
+		if len(args) != 1 {
+			return fmt.Errorf("USAGE: revoke path|serial")
+		}
+
+		v := connect()
+		return v.RevokeCertificate(args[0])
 	})
 
 	r.Dispatch("curl", func(command string, args ...string) error {